Moved ActivateStudent and InactivateStudent into the CourseMembership SS
