+++ app/controllers/api/v1/students_controller.rb
@@ -76,7 +76,7 @@ class Api::V1::StudentsController < Api::V1::ApiController
   EOS
   def destroy
     OSU::AccessPolicy.require_action_allowed!(:destroy, current_api_user, @student)
-    result = InactivateStudent.call(student: @student)
+    result = CourseMembership::InactivateStudent.call(student: @student)
 
     if result.errors.any?
       render_api_errors(result.errors)
@@ -93,7 +93,7 @@ class Api::V1::StudentsController < Api::V1::ApiController
   EOS
   def undrop
     OSU::AccessPolicy.require_action_allowed!(:destroy, current_api_user, @student)
-    result = ActivateStudent.call(student: @student)
+    result = CourseMembership::ActivateStudent.call(student: @student)
 
     if result.errors.any?
       render_api_errors(result.errors)
